refactor(genie): extract bootstrap-safe sync resolver so the closure walker is itself importable pre-install
The bootstrap-closure walker imported genie's resolver from core/import-map/mod.ts,
whose Effect functions eagerly pull @overeng/otel-contract (via observability). So
importing the walker required effect-utils' node deps — fine in effect-utils' own
post-install check:all, but a nix-packaged-genie downstream consumer (e.g. dotfiles)
that never installs those deps got `Cannot find module '@overeng/otel-contract'`.
Ironically the tool enforcing "importable pre-install" was not itself importable pre-install.

- Move the effect-free sync resolution path into core/import-map/sync-resolver.ts
  (only node:fs/node:path). mod.ts re-exports the public sync API unchanged.
- Point the walker at sync-resolver.ts. Proven: its static bundle now imports only
  typescript + node builtins — no effect/@effect/otel-contract/observability.
