+++ packages/@overeng/genie/docs/.decisions/0003-bootstrap-closure-engine.md
@@ -21,6 +21,7 @@ Sub-decisions:
 - **Policy is a strict allowlist**, not a runtime-only denylist: a violation is any bare specifier (non-relative, non-`#`/`#mr`, non-node-builtin), reusing the "external = non-relative, non-`node:`" definition already encoded by the `no-external-imports` rule. Node builtins are exempt (importable pre-install with or without the `node:` prefix). Bare first-party `@overeng/*` IS a violation (not resolvable pre-install; use relative or `#mr`).
 - **Enforcement is baseline + ratchet**: the pre-existing violations are baselined so the gate is green today; only new violations fail, and stale baseline entries are flagged so it ratchets down.
 - **Integration is a devenv task** (`bootstrap-closure:check`) wired into `check:all`, mirroring `weaver:check` — not a new genie CLI subcommand. The `checkBootstrapClosure` walker is exported from `@overeng/genie/node` for downstream reuse.
+- **The walker is itself bootstrap-safe.** It imports genie's resolver from `core/import-map/sync-resolver.ts` — the effect-free sync resolution path (only `node:fs`/`node:path`), extracted out of `import-map/mod.ts` (whose Effect functions eagerly pull `@overeng/otel-contract` via observability). So the walker's whole import closure is `typescript` + node builtins, and it runs in a nix-packaged-genie consumer that never installs effect-utils' node deps. `mod.ts` re-exports the sync API for backward compatibility. Without this, the tool that enforces "importable pre-install" is not itself importable pre-install (verified downstream in dotfiles).
 
 ## Options considered
